mdev57 5.5 main.file_contents fails on debian5-i386-fulltest.
The line in the file_contents.test removes all the '/lib' substrings from the
path, so file cannot be found if a path contains such a substring.
As i didn't find where it is needed, the line was just removed

per-file comments:
  mysql-test/t/file_contents.test
mdev57 5.5 main.file_contents fails on debian5-i386-fulltest.
        no '/lib' substring cutting.