Make RPMs for ULN build again.
A change to "vio/viosslfactories.c" in August, 2012,
broke a patch which is to be applied during the build
of ULN RPMs.
The patch file is
"packaging/rpm-uln/mysql-chain-certs.patch"

This change bypasses the problem by not trying to apply
the patch.

This is a regression and must be fixed, not bypassed.